+++ example/lib/main.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'dart:async';
@@ -7,6 +6,10 @@ import 'dart:io' show Platform, sleep;
 import 'package:flutter/services.dart';
 import 'package:flutter_nfc_kit/flutter_nfc_kit.dart';
 import 'package:ndef/ndef.dart' as ndef;
+
+import 'record-setting/raw_record_setting.dart';
+import 'record-setting/text_record_setting.dart';
+import 'record-setting/uri_record_setting.dart';
 
 void main() => runApp(MaterialApp(home: MyApp()));
 
@@ -128,359 +131,147 @@ class _MyAppState extends State<MyApp> with SingleTickerProviderStateMixin {
                     'ID: ${_tag?.id}\nStandard: ${_tag?.standard}\nType: ${_tag?.type}\nATQA: ${_tag?.atqa}\nSAK: ${_tag?.sak}\nHistorical Bytes: ${_tag?.historicalBytes}\nProtocol Info: ${_tag?.protocolInfo}\nApplication Data: ${_tag?.applicationData}\nHigher Layer Response: ${_tag?.hiLayerResponse}\nManufacturer: ${_tag?.manufacturer}\nSystem Code: ${_tag?.systemCode}\nDSF ID: ${_tag?.dsfId}\nNDEF Available: ${_tag?.ndefAvailable}\nNDEF Type: ${_tag?.ndefType}\nNDEF Writable: ${_tag?.ndefWritable}\nNDEF Can Make Read Only: ${_tag?.ndefCanMakeReadOnly}\nNDEF Capacity: ${_tag?.ndefCapacity}\n\n Transceive Result:\n$_result'),
               ])))),
           Center(
-            child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
+            child:
+                Column(mainAxisAlignment: MainAxisAlignment.start, children: <
+                    Widget>[
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceAround,
                 children: <Widget>[
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceAround,
-                    children: <Widget>[
-                      RaisedButton(
-                        onPressed: () async {
-                          if (_message.length != 0) {
-                            try {
-                              NFCTag tag = await FlutterNfcKit.poll();
-                              setState(() {
-                                _tag = tag;
-                              });
-                              if (tag.type == NFCTagType.mifare_ultralight ||
-                                  tag.type == NFCTagType.mifare_classic) {
-                                await FlutterNfcKit.writeNDEFRecords(_message);
-                                setState(() {
-                                  _writeResult = 'OK';
-                                });
-                              } else {
-                                setState(() {
-                                  _writeResult =
-                                      'error: NDEF not supported: ${tag.type}';
-                                });
-                              }
-                            } catch (e) {
-                              setState(() {
-                                _writeResult = 'error: $e';
-                              });
-                            }
+                  RaisedButton(
+                    onPressed: () async {
+                      if (_message.length != 0) {
+                        try {
+                          NFCTag tag = await FlutterNfcKit.poll();
+                          setState(() {
+                            _tag = tag;
+                          });
+                          if (tag.type == NFCTagType.mifare_ultralight ||
+                              tag.type == NFCTagType.mifare_classic) {
+                            await FlutterNfcKit.writeNDEFRecords(_message);
+                            setState(() {
+                              _writeResult = 'OK';
+                            });
                           } else {
                             setState(() {
-                              _writeResult = 'error: No record';
+                              _writeResult =
+                                  'error: NDEF not supported: ${tag.type}';
                             });
                           }
-                        },
-                        child: Text("Start writing"),
-                      ),
-                      RaisedButton(
-                        onPressed: () {
-                          showDialog(
-                            context: context,
-                            builder: (BuildContext context) {
-                              return SimpleDialog(
+                        } catch (e) {
+                          setState(() {
+                            _writeResult = 'error: $e';
+                          });
+                        } finally {
+                          await FlutterNfcKit.finish();
+                        }
+                      } else {
+                        setState(() {
+                          _writeResult = 'error: No record';
+                        });
+                      }
+                    },
+                    child: Text("Start writing"),
+                  ),
+                  RaisedButton(
+                    onPressed: () {
+                      showDialog(
+                          context: context,
+                          builder: (BuildContext context) {
+                            return SimpleDialog(
                                 title: Text("Record Type"),
                                 children: <Widget>[
                                   SimpleDialogOption(
                                     child: Text("Text Record"),
-                                    onPressed: () {
-                                      setState(() {
-                                        _message.add(ndef.TextRecord());
-                                      });
-                                      final result = await Navigator.push(context,
+                                    onPressed: () async {
+                                      final result = await Navigator.push(
+                                          context,
                                           MaterialPageRoute(builder: (context) {
-                                            return TextRecordSetting(
-                                              _message.last
-                                            );
-                                          })
-                                      );
+                                        return TextRecordSetting();
+                                      }));
+                                      if(result!=null) {
+                                        if (result is ndef.TextRecord) {
+                                          setState(() {
+                                            _message.add(result);
+                                          });
+                                        }
+                                      }
                                     },
                                   ),
                                   SimpleDialogOption(
                                     child: Text("Uri Record"),
-                                    onPressed: () {
-                                      setState(() {
-                                        _message.add(ndef.UriRecord());
-                                      });
-                                      final result = await Navigator.push(context,
+                                    onPressed: () async {
+                                      final result = await Navigator.push(
+                                          context,
                                           MaterialPageRoute(builder: (context) {
-                                            return UriRecordSetting(
-                                              _message.last
-                                            );
-                                          })
-                                      );
+                                        return UriRecordSetting();
+                                      }));
+                                      if(result!=null) {
+                                        if (result is ndef.UriRecord) {
+                                          setState(() {
+                                            _message.add(result);
+                                          });
+                                        }
+                                      }
                                     },
                                   ),
                                   SimpleDialogOption(
                                     child: Text("Raw Record"),
-                                    onPressed: () {
-                                      setState(() {
-                                        _message.add(ndef.UriRecord());
-                                      });
-                                      final result = await Navigator.push(context,
+                                    onPressed: () async {
+                                      final result = await Navigator.push(
+                                          context,
                                           MaterialPageRoute(builder: (context) {
-                                            return NDEFRecordSetting(
-                                              _message.last
-                                            );
-                                          })
-                                      );
+                                        return NDEFRecordSetting();
+                                      }));
+                                      if (result != null) {
+                                        if (result is ndef.NDEFRecord) {
+                                          setState(() {
+                                            _message.add(result);
+                                          });
+                                        }
+                                      }
                                     },
                                   ),
-                                ]
-                              );
-                            }
-                          );
-                        },
-                        child: Text("Add record"),
-                      )
-                    ],
-                  ),
-                  Text('Result:$_writeResult'),
-                  Expanded(
-                    flex: 1,
-                    child: ListView(
-                        shrinkWrap: true,
-                        children: List<Widget>.generate(
-                            _message.length,
-                            (index) => GestureDetector(
-                                  child: Text(
-                                      'id:${_message[index].id.toHexString()}\ntnf:${_message[index].tnf}\ntype:${_message[index].type.toHexString()}\npayload:${_message[index].payload.toHexString}\n'),
-                                  onTap: () async {
-                                    final result = await Navigator.push(context,
-                                        MaterialPageRoute(builder: (context) {
-                                      return NDEFRecordSetting(
-                                          record: _message[index]);
-                                    }));
-                                    if (result != null) {
-                                      if (result is ndef.NDEFRecord) {
-                                        setState(() {
-                                          _message[index] = result;
-                                        });
-                                      } else if (result is String &&
-                                          result == "Delete") {
-                                        _message.removeAt(index);
-                                      }
-                                    }
-                                  },
-                                ))),
-                  ),
-                ]),
+                                ]);
+                          });
+                    },
+                    child: Text("Add record"),
+                  )
+                ],
+              ),
+              Text('Result:$_writeResult'),
+              Expanded(
+                flex: 1,
+                child: ListView (
+                    shrinkWrap: true,
+                    children: List<Widget>.generate(
+                        _message.length,
+                        (index) => GestureDetector(
+                              child: Text(
+                                //id:${_message[index].id.toHexString()}\n
+                                  'tnf:${_message[index].tnf}\ntype:${_message[index].type.toHexString()}\npayload:${_message[index].payload.toHexString()}\n'),
+                              onTap: () async {
+                                final result = await Navigator.push(context,
+                                    MaterialPageRoute(builder: (context) {
+                                  return NDEFRecordSetting(
+                                      record: _message[index]);
+                                }));
+                                if (result != null) {
+                                  if (result is ndef.NDEFRecord) {
+                                    setState(() {
+                                      _message[index] = result;
+                                    });
+                                  } else if (result is String &&
+                                      result == "Delete") {
+                                    _message.removeAt(index);
+                                  }
+                                }
+                              },
+                            ))),
+              ),
+            ]),
           )
         ]),
       ),
     );
-  }
-}
-
-class NDEFRecordSetting extends StatefulWidget {
-  ndef.NDEFRecord record;
-  NDEFRecordSetting({Key key, this.record}) : super(key: key);
-  @override
-  _NDEFRecordSetting createState() => _NDEFRecordSetting();
-}
-
-class _NDEFRecordSetting extends State<NDEFRecordSetting> {
-  GlobalKey _formKey = new GlobalKey<FormState>();
-  TextEditingController _identifierController;
-  TextEditingController _payloadController;
-  TextEditingController _typeController;
-  int _dropButtonValue;
-
-  @override
-  initState() {
-    _identifierController = new TextEditingController.fromValue(
-        TextEditingValue(text: widget.record.id.toHexString()));
-    _payloadController = new TextEditingController.fromValue(
-        TextEditingValue(text: widget.record.payload.toHexString()));
-    _typeController = new TextEditingController.fromValue(
-        TextEditingValue(text: widget.record.type.toHexString()));
-    _dropButtonValue = ndef.TypeNameFormat.values.indexOf(widget.record.tnf);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        home: Scaffold(
-            appBar: AppBar(
-              title: Text('Set Record'),
-            ),
-            body: Center(
-                child: Form(
-                    key: _formKey,
-                    autovalidate: true,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      children: <Widget>[
-                        DropdownButton(
-                          value: _dropButtonValue,
-                          items: [
-                            DropdownMenuItem(
-                              child: Text('empty'),
-                              value: 0,
-                            ),
-                            DropdownMenuItem(
-                              child: Text('nfcWellKnown'),
-                              value: 1,
-                            ),
-                            DropdownMenuItem(
-                              child: Text('media'),
-                              value: 2,
-                            ),
-                            DropdownMenuItem(
-                              child: Text('absoluteURI'),
-                              value: 3,
-                            ),
-                            DropdownMenuItem(
-                                child: Text('nfcExternal'), value: 4),
-                            DropdownMenuItem(
-                                child: Text('unchanged'), value: 5),
-                            DropdownMenuItem(
-                              child: Text('unknown'),
-                              value: 6,
-                            ),
-                          ],
-                          onChanged: (value) {
-                            setState(() {
-                              _dropButtonValue = value;
-                            });
-                          },
-                        ),
-                        TextFormField(
-                          decoration: InputDecoration(labelText: 'identifier'),
-                          validator: (v) {
-                            return v.trim().length % 2 == 0
-                                ? null
-                                : 'length must be even';
-                          },
-                          controller: _identifierController,
-                        ),
-                        TextFormField(
-                          decoration: InputDecoration(labelText: 'type'),
-                          validator: (v) {
-                            return v.trim().length % 2 == 0
-                                ? null
-                                : 'length must be even';
-                          },
-                          controller: _typeController,
-                        ),
-                        TextFormField(
-                          decoration: InputDecoration(labelText: 'payload'),
-                          validator: (v) {
-                            return v.trim().length % 2 == 0
-                                ? null
-                                : 'length must be even';
-                          },
-                          controller: _payloadController,
-                        ),
-                        RaisedButton(
-                          child: Text('OK'),
-                          onPressed: () {
-                            if ((_formKey.currentState as FormState)
-                                .validate()) {
-                              Navigator.pop(
-                                  context,
-                                  ndef.NDEFRecord(tnf:_dropButtonValue,type:(_typeController.text==null?"":_typeController.text).toBytes(),id:(_identifierController.text==null?"":_identifierController.text).toBytes(),payload:(_payloadController.text==null?"":_payloadController.text).toBytes());
-                            }
-                          },
-                        ),
-                        RaisedButton(
-                          child: Text('Delete'),
-                          onPressed: () {
-                            if ((_formKey.currentState as FormState)
-                                .validate()) {
-                              Navigator.pop(context, 'Delete');
-                            }
-                          },
-                        ),
-                      ],
-                    )))));
-  }
-}
-
-class TextRecordSetting extends StatefulWidget {
-  ndef.TextRecord record;
-  TextRecordSetting({Key key, this.record}) : super(key: key);
-  @override
-  _TextRecordSetting createState() => _TextRecordSetting();
-}
-
-class _TextRecordSetting extends State<TextRecordSetting> {
-  GlobalKey _formKey = new GlobalKey<FormState>();
-  TextEditingController _languageController;
-  TextEditingController _textController;
-  int _dropButtonValue;
-
-  @override
-  initState() {
-    _languageController = new TextEditingController.fromValue(
-        TextEditingValue(text: widget.record.language));
-    _textController = new TextEditingController.fromValue(
-        TextEditingValue(text: widget.record.text));
-    _dropButtonValue = ndef.TextEncoding.values.indexOf(widget.record.encoding);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        home: Scaffold(
-            appBar: AppBar(
-              title: Text('Set Record'),
-            ),
-            body: Center(
-                child: Form(
-                    key: _formKey,
-                    autovalidate: true,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      children: <Widget>[
-                        DropdownButton(
-                          value: _dropButtonValue,
-                          items: [
-                            DropdownMenuItem(
-                                child: Text('UTF-8'), value: 0),
-                            DropdownMenuItem(
-                                child: Text('UTF-16'), value: 1),
-                          ],
-                          onChanged: (value) {
-                            setState(() {
-                              _dropButtonValue = value;
-                            });
-                          },
-                        ),
-                        TextFormField(
-                          decoration: InputDecoration(labelText: 'language'),
-                          validator: (v) {
-                            return v.trim().length % 2 == 0
-                                ? null
-                                : 'length must not be blank';
-                          },
-                          controller: _languageController,
-                        ),
-                        TextFormField(
-                          decoration: InputDecoration(labelText: 'text'),
-                          controller: _textController,
-                        ),
-                        RaisedButton(
-                          child: Text('OK'),
-                          onPressed: () {
-                            if ((_formKey.currentState as FormState)
-                                .validate()) {
-                              Navigator.pop(
-                                  context,
-                                  ndef.TextRecord(encoding: ndef.TextEncoding.values[_dropButtonValue],
-                                      language: (_languageController.text == null
-                                      ? ""
-                                      : _languageController.text),
-                                      text: (_textController.text == null
-                                      ? ""
-                                      : _textController.text)));
-                            }
-                          },
-                        ),
-                        RaisedButton(
-                          child: Text('Delete'),
-                          onPressed: () {
-                            if ((_formKey.currentState as FormState)
-                                .validate()) {
-                              Navigator.pop(context, 'Delete');
-                            }
-                          },
-                        ),
-                      ],
-                    )))));
   }
 }