+++ example/lib/record-setting/text_record_setting.dart
@@ -1,0 +1,95 @@
+import 'package:flutter/material.dart';
+
+import 'package:ndef/ndef.dart' as ndef;
+
+class TextRecordSetting extends StatefulWidget {
+  ndef.TextRecord record;
+  TextRecordSetting({Key key, ndef.TextRecord record}) : super(key: key) {
+    if(record==null) {
+      this.record = ndef.TextRecord(language: 'en', text: '');
+    } else {
+      this.record = record;
+    }
+  }
+  @override
+  _TextRecordSetting createState() => _TextRecordSetting();
+}
+
+class _TextRecordSetting extends State<TextRecordSetting> {
+  GlobalKey _formKey = new GlobalKey<FormState>();
+  TextEditingController _languageController;
+  TextEditingController _textController;
+  int _dropButtonValue;
+
+  @override
+  initState() {
+    _languageController = new TextEditingController.fromValue(
+        TextEditingValue(text: widget.record.language));
+    _textController = new TextEditingController.fromValue(
+        TextEditingValue(text: widget.record.text));
+    _dropButtonValue = ndef.TextEncoding.values.indexOf(widget.record.encoding);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+        home: Scaffold(
+            appBar: AppBar(
+              title: Text('Set Record'),
+            ),
+            body: Center(
+                child: Form(
+                    key: _formKey,
+                    autovalidate: true,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: <Widget>[
+                        DropdownButton(
+                          value: _dropButtonValue,
+                          items: [
+                            DropdownMenuItem(child: Text('UTF-8'), value: 0),
+                            DropdownMenuItem(child: Text('UTF-16'), value: 1),
+                          ],
+                          onChanged: (value) {
+                            setState(() {
+                              _dropButtonValue = value;
+                            });
+                          },
+                        ),
+                        TextFormField(
+                          decoration: InputDecoration(labelText: 'language'),
+                          validator: (v) {
+                            return v.trim().length % 2 == 0
+                                ? null
+                                : 'length must not be blank';
+                          },
+                          controller: _languageController,
+                        ),
+                        TextFormField(
+                          decoration: InputDecoration(labelText: 'text'),
+                          controller: _textController,
+                        ),
+                        RaisedButton(
+                          child: Text('OK'),
+                          onPressed: () {
+                            if ((_formKey.currentState as FormState)
+                                .validate()) {
+                              Navigator.pop(
+                                  context,
+                                  ndef.TextRecord(
+                                      encoding: ndef.TextEncoding
+                                          .values[_dropButtonValue],
+                                      language:
+                                          (_languageController.text == null
+                                              ? ""
+                                              : _languageController.text),
+                                      text: (_textController.text == null
+                                          ? ""
+                                          : _textController.text)));
+                            }
+                          },
+                        ),
+                      ],
+                    )))));
+  }
+}